Actualizacion de casting.py

+++ casting.py
@@ -1,6 +1,6 @@
 print("ingresa el primer punto")
 num1 = int(input())
 print("ingresa el segundo punto")
-num2 = int(input())
+num2 = float(input())
 resultado = num1 + num2 
 print("El resultado es:" + str(resultado))
